#89: 버튼 생성 함수 추가
- UIButton을 상속하는 대신 이미지가 설정된 버튼을 반환하는 함수를 추가하였습니다.
- 관련 코드리뷰 링크: https://github.com/ToDo-Garden/ToDo-Garden/pull/90#pullrequestreview-2044669121
